requirements.txt
langgraph>=0.0.19
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.0.11
openai>=1.12.0
anthropic>=0.8.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0 